Create tests for the new PseudoField
